zza: simplified server
